Abstract the text based PartitionReader
